+++ src/pipcanary/module_loader.py
@@ -1,22 +1,25 @@
+import os
 import sys
 import pkgutil
-
-from distutils.sysconfig import get_python_lib
+import importlib
 
 
 def load_modules(dirname):
-    for importer, package_name, _ in pkgutil.iter_modules([dirname]):
+    for _, package_name, _ in pkgutil.iter_modules([dirname]):
         if package_name not in sys.modules:
             try:
                 print("Loading package: %s..." % package_name)
                 print("Package: %s" % package_name, file=sys.stderr)
-                importer.find_module(package_name).load_module(  # type: ignore
-                    package_name
-                )
+                module = importlib.import_module(str(package_name))
+                print("Module: %s" % module)
                 del sys.modules[package_name]
-            except ImportError:
-                pass
+            except Exception as e:
+                print("Info: %s" % str(e))
 
 
 if __name__ == "__main__":
-    load_modules(get_python_lib())
+    filedir = os.path.dirname(__file__)
+    version = sys.version_info
+    libdir = f"{filedir}/lib/python{version.major}.{version.minor}/site-packages"
+    print("Loading modules from: %s..." % libdir)
+    load_modules(libdir)